refactor: move TUI files to presentation root
Move TUI implementation files from presentation/tui/ to presentation/
directory and remove unnecessary tui subdirectory. Update package name
from 'tui' to 'presentation' and fix all import references.
