+++ internal/presentation/view.go
@@ -1,4 +1,4 @@
-package tui
+package presentation
 
 import (
 	"fmt"
