+++ internal/presentation/view_test.go
@@ -1,4 +1,4 @@
-package tui
+package presentation
 
 import (
 	"testing"
